Adjust wolf step height and only follow owner while unsaddled.
Step height of 1.6 is a little too much, 1.4 seems to be sufficient, but
1.1 isn't quite enough to overcome all of the blocks in the Atum2 dimension.

However, it's speed will result in a lot of "Entity Desert Wolf moved
wrongly!" messages being spammed in chat. It has to do with the speed at
which something moves vertically (i.e., with step assist). Other than
the warning, it should have no other effect. Reducing the overall speed
of the mob may help.

I've only had that issue with the Botania skelefied horses/zombified
horses.

The new EntityAIFollowOwnerWithoutSaddle task literally derives from
EntityAIFollowOwner, and simply overrides should execute to fail early
if the entity is saddled.